Add yarn.lock file
We are using yarn to manage our dependencies directly, so we need to add
the yarn.lock file to the repo so everyone gets the same versions of
dependencides.

Yarn is a very useful tool that allows for caching dependencies and
locking the versions so that all developers are using the same versions
during development.